refactor: 기능별 코드 분리
